refactor(py): consolidate coercion helpers
Closes THE-2443
